Added the handle_userInput function
This uses basic error handling to determine what room the character will move to based on the provided users input. If no room exists, then it will tell them so and it what direction they tried to go in. If the user typed in an invalid option I have the script prompt the user to type in a specific letter