MVP finally met
Rooms can individually hold items. Player can now pick up and drop items, and it adds it back to the room. Added the function has_items to the loop to check if the room the player is in has items, the add function just wasn't working for me. Adjusted some of the formating for the REPL so it was easier to read on both input and output