MVP finally met
Removed the display items call here as it was not working. IT would print out all items in any room, in every room upon entry. In adv.py I found a solution fix for this. I adjusted the REPL input and output formatting so it appears easier to read and is formatted better.